splitting partition parameter builder into histogram and value counts parameter builders
