+++ great_expectations/rule_based_profiler/data_assistant/data_assistant.py
@@ -29,11 +29,11 @@ from great_expectations.rule_based_profiler.helpers.runtime_environment import (
 )
 from great_expectations.rule_based_profiler.helpers.util import sanitize_parameter_name
 from great_expectations.rule_based_profiler.parameter_builder import (
+    HistogramSingleBatchParameterBuilder,
     MeanUnexpectedMapMetricMultiBatchParameterBuilder,
     MetricMultiBatchParameterBuilder,
     NumericMetricRangeMultiBatchParameterBuilder,
     ParameterBuilder,
-    PartitionParameterBuilder,
 )
 from great_expectations.rule_based_profiler.parameter_builder.regex_pattern_string_parameter_builder import (
     RegexPatternStringParameterBuilder,
@@ -361,17 +361,15 @@ class DataAssistant(metaclass=MetaDataAssistant):
             )
 
         @staticmethod
-        def build_partition_parameter_builder(
+        def build_histogram_single_batch_parameter_builder(
             name: str,
-            bucketize_data: Union[str, bool] = True,
-        ) -> PartitionParameterBuilder:
+        ) -> HistogramSingleBatchParameterBuilder:
             """
-            This method instantiates "PartitionParameterBuilder" class with specific arguments for given purpose.
+            This method instantiates "HistogramSingleBatchParameterBuilder" class with specific arguments for given purpose.
             """
             name: str = sanitize_parameter_name(name=f"{name}")
-            return PartitionParameterBuilder(
+            return HistogramSingleBatchParameterBuilder(
                 name=name,
-                bucketize_data=bucketize_data,
                 evaluation_parameter_builder_configs=None,
                 data_context=None,
             )
