+++ great_expectations/rule_based_profiler/data_assistant/onboarding_data_assistant.py
@@ -302,9 +302,8 @@ class OnboardingDataAssistant(DataAssistant):
 
         # Step-2: Declare "ParameterBuilder" for every metric of interest.
 
-        column_partition_parameter_builder_for_metrics: ParameterBuilder = DataAssistant.commonly_used_parameter_builders.build_partition_parameter_builder(
+        column_histogram_single_batch_parameter_builder_for_metrics: ParameterBuilder = DataAssistant.commonly_used_parameter_builders.build_histogram_single_batch_parameter_builder(
             name="column_values.partition",
-            bucketize_data=True,
         )
         column_min_metric_multi_batch_parameter_builder_for_metrics: ParameterBuilder = (
             DataAssistant.commonly_used_parameter_builders.get_column_min_metric_multi_batch_parameter_builder()
@@ -563,7 +562,7 @@ class OnboardingDataAssistant(DataAssistant):
             "round_decimals": 15,
         }
         parameter_builders: List[ParameterBuilder] = [
-            column_partition_parameter_builder_for_metrics,
+            column_histogram_single_batch_parameter_builder_for_metrics,
             column_min_metric_multi_batch_parameter_builder_for_metrics,
             column_max_metric_multi_batch_parameter_builder_for_metrics,
             column_quantile_values_metric_multi_batch_parameter_builder_for_metrics,
