+++ great_expectations/rule_based_profiler/parameter_builder/histogram_single_batch_parameter_builder.py
@@ -1,0 +1,241 @@
+from typing import Any, Dict, List, Optional, Set
+
+import numpy as np
+
+import great_expectations.exceptions as ge_exceptions
+from great_expectations.rule_based_profiler.config import ParameterBuilderConfig
+from great_expectations.rule_based_profiler.domain import Domain
+from great_expectations.rule_based_profiler.helpers.util import (
+    NP_EPSILON,
+    get_parameter_value_and_validate_return_type,
+)
+from great_expectations.rule_based_profiler.metric_computation_result import MetricValue
+from great_expectations.rule_based_profiler.parameter_builder import (
+    MetricSingleBatchParameterBuilder,
+)
+from great_expectations.rule_based_profiler.parameter_container import (
+    DOMAIN_KWARGS_PARAMETER_FULLY_QUALIFIED_NAME,
+    FULLY_QUALIFIED_PARAMETER_NAME_METADATA_KEY,
+    FULLY_QUALIFIED_PARAMETER_NAME_VALUE_KEY,
+    RAW_PARAMETER_KEY,
+    ParameterContainer,
+    ParameterNode,
+)
+from great_expectations.types.attributes import Attributes
+from great_expectations.util import (
+    convert_ndarray_datetime_to_float_dtype,
+    is_ndarray_datetime_dtype,
+)
+
+
+class HistogramSingleBatchParameterBuilder(MetricSingleBatchParameterBuilder):
+    """
+    Compute histogram using specified metric for one Batch of data.
+    """
+
+    exclude_field_names: Set[
+        str
+    ] = MetricSingleBatchParameterBuilder.exclude_field_names | {
+        "column_partition_metric_single_batch_parameter_builder_config",
+        "metric_name",
+        "metric_domain_kwargs",
+        "metric_value_kwargs",
+        "enforce_numeric_metric",
+        "replace_nan_with_zero",
+        "reduce_scalar_metric",
+    }
+
+    def __init__(
+        self,
+        name: str,
+        bins: str = "uniform",
+        n_bins: int = 10,
+        allow_relative_error: bool = False,
+        evaluation_parameter_builder_configs: Optional[
+            List[ParameterBuilderConfig]
+        ] = None,
+        data_context: Optional["BaseDataContext"] = None,  # noqa: F821
+    ) -> None:
+        """
+        Args:
+            name: the name of this parameter -- this is user-specified parameter name (from configuration);
+            it is not the fully-qualified parameter name; a fully-qualified parameter name must start with "$parameter."
+            and may contain one or more subsequent parts (e.g., "$parameter.<my_param_from_config>.<metric_name>").
+            bins: Partitioning strategy (one of "uniform", "ntile", "quantile", "percentile", or "auto"); please refer
+            to "ColumnPartition" (great_expectations/expectations/metrics/column_aggregate_metrics/column_partition.py).
+            n_bins: Number of bins for histogram computation (ignored and recomputed if "bins" argument is "auto").
+            allow_relative_error: Used for partitionong strategy values that involve quantiles (all except "uniform").
+            evaluation_parameter_builder_configs: ParameterBuilder configurations, executing and making whose respective
+            ParameterBuilder objects' outputs available (as fully-qualified parameter names) is pre-requisite.
+            These "ParameterBuilder" configurations help build parameters needed for this "ParameterBuilder".
+            data_context: BaseDataContext associated with this ParameterBuilder
+        """
+
+        self._column_partition_metric_single_batch_parameter_builder_config = (
+            ParameterBuilderConfig(
+                module_name="great_expectations.rule_based_profiler.parameter_builder",
+                class_name="MetricSingleBatchParameterBuilder",
+                name="column_partition_metric_single_batch_parameter_builder",
+                metric_name="column.partition",
+                metric_domain_kwargs=DOMAIN_KWARGS_PARAMETER_FULLY_QUALIFIED_NAME,
+                metric_value_kwargs={
+                    "bins": bins,
+                    "n_bins": n_bins,
+                    "allow_relative_error": allow_relative_error,
+                },
+                enforce_numeric_metric=False,
+                replace_nan_with_zero=False,
+                reduce_scalar_metric=False,
+                evaluation_parameter_builder_configs=None,
+            )
+        )
+
+        if evaluation_parameter_builder_configs is None:
+            evaluation_parameter_builder_configs = [
+                self._column_partition_metric_single_batch_parameter_builder_config,
+            ]
+
+        super().__init__(
+            name=name,
+            metric_name=None,
+            metric_domain_kwargs=DOMAIN_KWARGS_PARAMETER_FULLY_QUALIFIED_NAME,
+            metric_value_kwargs=None,
+            enforce_numeric_metric=False,
+            replace_nan_with_zero=False,
+            reduce_scalar_metric=False,
+            evaluation_parameter_builder_configs=evaluation_parameter_builder_configs,
+            data_context=data_context,
+        )
+
+    def _build_parameters(
+        self,
+        domain: Domain,
+        variables: Optional[ParameterContainer] = None,
+        parameters: Optional[Dict[str, ParameterContainer]] = None,
+        recompute_existing_parameter_values: bool = False,
+    ) -> Attributes:
+        """
+        Builds ParameterContainer object that holds ParameterNode objects with attribute name-value pairs and details.
+
+        Returns:
+            Attributes object, containing computed parameter values and parameter computation details metadata.
+        """
+        fully_qualified_column_partition_metric_single_batch_parameter_builder_name: str = f"{RAW_PARAMETER_KEY}{self._column_partition_metric_single_batch_parameter_builder_config.name}"
+        # Obtain "column.partition" from "rule state" (i.e., variables and parameters); from instance variable otherwise.
+        column_partition_parameter_node: ParameterNode = get_parameter_value_and_validate_return_type(
+            domain=domain,
+            parameter_reference=fully_qualified_column_partition_metric_single_batch_parameter_builder_name,
+            expected_return_type=None,
+            variables=variables,
+            parameters=parameters,
+        )
+        bins: MetricValue = column_partition_parameter_node[
+            FULLY_QUALIFIED_PARAMETER_NAME_VALUE_KEY
+        ]
+
+        element: Any
+        if bins is None or all(element is None for element in bins):
+            raise ge_exceptions.ProfilerExecutionError(
+                message=f"""Partitioning values for {self.__class__.__name__} by \
+{self._column_partition_metric_single_batch_parameter_builder_config.name} into bins encountered empty or non-existent \
+elements.
+"""
+            )
+
+        ndarray_is_datetime_type: bool = is_ndarray_datetime_dtype(
+            data=bins,
+            parse_strings_as_datetimes=True,
+        )
+        bins_ndarray_as_float: MetricValue
+        if ndarray_is_datetime_type:
+            bins_ndarray_as_float = convert_ndarray_datetime_to_float_dtype(data=bins)
+        else:
+            bins_ndarray_as_float = bins
+
+        histogram_unsupported: bool = ndarray_is_datetime_type or not np.all(
+            np.diff(bins_ndarray_as_float) > 0.0
+        )
+        if histogram_unsupported:
+            raise ge_exceptions.ProfilerExecutionError(
+                message=f"""Partitioning values for {self.__class__.__name__} by \
+{self._column_partition_metric_single_batch_parameter_builder_config.name} did not yield monotonically-increasing bins \
+of supported data type.
+"""
+            )
+
+        column_values_nonnull_count_metric_single_batch_parameter_builder = MetricSingleBatchParameterBuilder(
+            name="column_values_nonnull_count_metric_single_batch_parameter_builder",
+            metric_name="column_values.nonnull.count",
+            metric_domain_kwargs=DOMAIN_KWARGS_PARAMETER_FULLY_QUALIFIED_NAME,
+            metric_value_kwargs=None,
+            enforce_numeric_metric=False,
+            replace_nan_with_zero=False,
+            reduce_scalar_metric=False,
+            evaluation_parameter_builder_configs=None,
+            data_context=self.data_context,
+        )
+        column_values_nonnull_count_metric_single_batch_parameter_builder.build_parameters(
+            domain=domain,
+            variables=variables,
+            parameters=parameters,
+            batch_list=self.batch_list,
+            batch_request=self.batch_request,
+            recompute_existing_parameter_values=recompute_existing_parameter_values,
+        )
+        # Obtain "column_values.nonnull.count" from "rule state" (i.e., variables and parameters); from instance variable otherwise.
+        column_values_nonnull_count_parameter_node: ParameterNode = get_parameter_value_and_validate_return_type(
+            domain=domain,
+            parameter_reference=column_values_nonnull_count_metric_single_batch_parameter_builder.raw_fully_qualified_parameter_name,
+            expected_return_type=None,
+            variables=variables,
+            parameters=parameters,
+        )
+
+        self.metric_name = "column.histogram"
+        self.metric_value_kwargs = {
+            "bins": tuple(bins),
+        }
+
+        # Compute metric value for one Batch object.
+        super().build_parameters(
+            domain=domain,
+            variables=variables,
+            parameters=parameters,
+            parameter_computation_impl=super()._build_parameters,
+            recompute_existing_parameter_values=recompute_existing_parameter_values,
+        )
+
+        # Retrieve metric values for one Batch object.
+        parameter_node: ParameterNode = get_parameter_value_and_validate_return_type(
+            domain=domain,
+            parameter_reference=self.raw_fully_qualified_parameter_name,
+            expected_return_type=None,
+            variables=variables,
+            parameters=parameters,
+        )
+
+        # in this case, we have requested a partition, histogram using said partition, and nonnull count
+        bins = list(bins)
+        weights: np.ndarray = np.asarray(
+            parameter_node[FULLY_QUALIFIED_PARAMETER_NAME_VALUE_KEY]
+        ) / (
+            column_values_nonnull_count_parameter_node[
+                FULLY_QUALIFIED_PARAMETER_NAME_VALUE_KEY
+            ]
+            + NP_EPSILON
+        )
+        tail_weights: float = (1.0 - sum(weights)) / 2.0
+
+        partition_object: dict = {
+            "bins": bins,
+            "weights": weights.tolist(),
+            "tail_weights": [tail_weights, tail_weights],
+        }
+        details: dict = parameter_node[FULLY_QUALIFIED_PARAMETER_NAME_METADATA_KEY]
+
+        return Attributes(
+            {
+                FULLY_QUALIFIED_PARAMETER_NAME_VALUE_KEY: partition_object,
+                FULLY_QUALIFIED_PARAMETER_NAME_METADATA_KEY: details,
+            }
+        )
