+++ tests/rule_based_profiler/domain_builder/test_domain_builder.py
@@ -349,7 +349,7 @@ def test_column_pair_domain_builder_wrong_column_names(
     )
 
     with pytest.raises(ge_exceptions.ProfilerExecutionError) as excinfo:
-        # noinspection PyArgumentList
+        # noinspection PyUnusedLocal
         domains: List[Domain] = domain_builder.get_domains(
             rule_name="my_rule", variables=variables, batch_request=batch_request
         )
@@ -456,7 +456,7 @@ def test_multi_column_domain_builder_wrong_column_list(
     )
 
     with pytest.raises(ge_exceptions.ProfilerExecutionError) as excinfo:
-        # noinspection PyArgumentList
+        # noinspection PyUnusedLocal
         domains: List[Domain] = domain_builder.get_domains(
             rule_name="my_rule", variables=variables, batch_request=batch_request
         )
@@ -466,7 +466,7 @@ def test_multi_column_domain_builder_wrong_column_list(
     )
 
     with pytest.raises(ge_exceptions.ProfilerExecutionError) as excinfo:
-        # noinspection PyArgumentList
+        # noinspection PyUnusedLocal
         domains: List[Domain] = domain_builder.get_domains(
             rule_name="my_rule", variables=variables, batch_request=batch_request
         )
